Move conditional opportunity pieces into a separate file
